Updated CMakeLists.txt, now the game can be run from VS.

+++ src/Drawing.cpp
@@ -63,8 +63,8 @@ namespace draw
             GLfloat texLeft = 0.f;
             GLfloat texRight = (GLfloat)imgWidth / (GLfloat)texWidth;
 
-            GLfloat quadWidth = imgWidth;
-            GLfloat quadHeight = imgHeight;
+            GLfloat quadWidth = (GLfloat)imgWidth;
+            GLfloat quadHeight = (GLfloat)imgHeight;
 
             if(clip != NULL) 
             {
@@ -116,7 +116,7 @@ namespace draw
 
         GLfloat curX = x;
         GLfloat curY = y;
-        for(int i = 0; i < text.size(); ++i)
+        for(int i = 0; i < (int)text.size(); ++i)
         {
             if(text[i] == '\n' || text[i] == '\r')
             {
